+++ cucumber-core/src/test/java/io/cucumber/core/backend/StubHookDefinition.java
@@ -6,49 +6,46 @@ import java.util.function.Consumer;
 public class StubHookDefinition implements HookDefinition {
 
     private static final String STUBBED_LOCATION_WITH_DETAILS = "{stubbed location with details}";
-    private final String location;
+    private final Located location;
     private final RuntimeException exception;
     private final Consumer<TestCaseState> action;
-    private final SourceReference sourceReference;
     private final HookType hookType;
 
     public StubHookDefinition(
-            String location, RuntimeException exception, Consumer<TestCaseState> action,
-            SourceReference sourceReference, HookType hookType
+            Located location, RuntimeException exception, Consumer<TestCaseState> action, HookType hookType
     ) {
         this.location = location;
         this.exception = exception;
         this.action = action;
-        this.sourceReference = sourceReference;
         this.hookType = hookType;
     }
 
-    public StubHookDefinition(String location, Consumer<TestCaseState> action) {
-        this(location, null, action, null, null);
+    public StubHookDefinition(String location) {
+        this(new StubLocation(location), null, null, null);
+    }
+
+    public StubHookDefinition(SourceReference location, HookType hookType, Consumer<TestCaseState> action) {
+        this(new StubLocation(location), null, action, hookType);
     }
 
     public StubHookDefinition() {
-        this(STUBBED_LOCATION_WITH_DETAILS, null, null, null, null);
+        this(new StubLocation(STUBBED_LOCATION_WITH_DETAILS), null, null, null);
     }
 
     public StubHookDefinition(Consumer<TestCaseState> action) {
-        this(STUBBED_LOCATION_WITH_DETAILS, null, action, null, null);
+        this(new StubLocation(STUBBED_LOCATION_WITH_DETAILS), null, action, null);
     }
 
     public StubHookDefinition(RuntimeException exception) {
-        this(STUBBED_LOCATION_WITH_DETAILS, exception, null, null, null);
-    }
-
-    public StubHookDefinition(String location) {
-        this(location, null, null, null, null);
+        this(new StubLocation(STUBBED_LOCATION_WITH_DETAILS), exception, null, null);
     }
 
     public StubHookDefinition(SourceReference sourceReference, HookType hookType) {
-        this(null, null, null, sourceReference, hookType);
+        this(new StubLocation(sourceReference), null, null, hookType);
     }
 
     public StubHookDefinition(SourceReference sourceReference, HookType hookType, RuntimeException exception) {
-        this(null, exception, null, sourceReference, hookType);
+        this(new StubLocation(sourceReference), exception, null, hookType);
     }
 
     @Override
@@ -78,12 +75,12 @@ public class StubHookDefinition implements HookDefinition {
 
     @Override
     public String getLocation() {
-        return location;
+        return location.getLocation();
     }
 
     @Override
     public Optional<SourceReference> getSourceReference() {
-        return Optional.ofNullable(sourceReference);
+        return location.getSourceReference();
     }
 
     @Override