+++ cucumber-core/src/test/java/io/cucumber/core/plugin/JsonFormatterTestStepDefinitions.java
@@ -1,0 +1,83 @@
+package io.cucumber.core.plugin;
+
+class JsonFormatterTestStepDefinitions {
+
+    public void bg_1() {
+    }
+
+    public void bg_2() {
+    }
+
+    public void bg_3() {
+    }
+
+    public void step_1() {
+    }
+
+    public void step_2() {
+    }
+
+    public void step_3() {
+    }
+
+    public void cliche() {
+    }
+
+    public void so_1() {
+    }
+
+    public void so_2() {
+    }
+
+    public void so_3() {
+    }
+
+    public void a() {
+    }
+
+    public void b() {
+    }
+
+    public void c() {
+    }
+
+    public void before_hook_1() {
+
+    }
+
+    public void after_hook_1() {
+
+    }
+
+    public void beforestep_hook_1() {
+
+    }
+
+    public void afterstep_hook_1() {
+
+    }
+
+    public void afterstep_hook_2() {
+
+    }
+
+    public void there_are_bananas() {
+
+    }
+
+    public void there_are_oranges() {
+
+    }
+
+    public void monkey_eats_bananas() {
+
+    }
+
+    public void monkey_eats_more_bananas() {
+
+    }
+
+    public void monkey_arrives() {
+
+    }
+}